Function pangocairo::show_error_underline

pub fn show_error_underline(
    cr: &Context, 
    x: f64, 
    y: f64, 
    width: f64, 
    height: f64
)

Draw a squiggly line in the specified cairo context that approximately covers the given rectangle in the style of an underline used to indicate a spelling error. (The width of the underline is rounded to an integer number of up/down segments and the resulting rectangle is centered in the original rectangle)

cr

a Cairo context

x

The X coordinate of one corner of the rectangle

y

The Y coordinate of one corner of the rectangle

width

Non-negative width of the rectangle

height

Non-negative height of the rectangle