Function pangocairo::show_glyph_string

pub fn show_glyph_string<P: IsA<Font>>(
    cr: &Context, 
    font: &P, 
    glyphs: &mut GlyphString
)

Draws the glyphs in glyphs in the specified cairo context. The origin of the glyphs (the left edge of the baseline) will be drawn at the current point of the cairo context.

cr

a Cairo context

font

a pango::Font from a FontMap

glyphs

a pango::GlyphString