pangocairo::functions

Function context_set_font_options

pub fn context_set_font_options(
    context: &Context,
    options: Option<&FontOptions>,
)

Sets the font options used when rendering text with this context.

These options override any options that [update_context()][crate::update_context()] derives from the target surface.

context

a pango::Context, from a pangocairo font map

options

a cairo::FontOptions, or None to unset any previously set options. A copy is made.