pangocairo::functions

Function layout_line_path

pub fn layout_line_path(cr: &Context, line: &LayoutLine)

Adds the text in pango::LayoutLine to the current path in the specified cairo context.

The origin of the glyphs (the left edge of the line) will be at the current point of the cairo context.

cr

a Cairo context

line

a pango::LayoutLine