Function pangocairo::functions::layout_path

pub fn layout_path(cr: &Context, layout: &Layout)

Adds the text in a pango::Layout to the current path in the specified cairo context.

The top-left corner of the pango::Layout will be at the current point of the cairo context.

cr

a Cairo context

layout

a Pango layout