pangocairo::functions
Function show_glyph_item

pub fn show_glyph_item(cr: &Context, text: &str, glyph_item: &mut GlyphItem)

Draws the glyphs in @glyph_item in the specified cairo context,

embedding the text associated with the glyphs in the output if the output format supports it (PDF for example), otherwise it acts similar to [show_glyph_string()][crate::show_glyph_string()].

The origin of the glyphs (the left edge of the baseline) will be drawn at the current point of the cairo context.

Note that @text is the start of the text for layout, which is then indexed by glyph_item->item->offset.

cr

a Cairo context

text

the UTF-8 text that @glyph_item refers to

glyph_item

a pango::GlyphItem