pangocairo::functions

Function update_context

pub fn update_context(cr: &Context, context: &Context)

Updates a pango::Context previously created for use with Cairo to match the current transformation and target surface of a Cairo context.

If any layouts have been created for the context, it’s necessary to call Layout::context_changed() on those layouts.

cr

a Cairo context

context

a pango::Context, from a pangocairo font map