Function pangocairo::prelude::functions::create_layout

pub fn create_layout(cr: &Context) -> Option<Layout>

Creates a layout object set up to match the current transformation and target surface of the Cairo context. This layout can then be used for text measurement with functions like [crate::pango::Layout::get_size()] or drawing with functions like pango_cairo_show_layout(). If you change the transformation or target surface for cr, you need to call pango_cairo_update_layout()

This function is the most convenient way to use Cairo with Pango, however it is slightly inefficient since it creates a separate crate::pango::Context object for each layout. This might matter in an application that was laying out large amounts of text.

cr

a Cairo context

Returns

the newly created crate::pango::Layout. Free with [crate::glib::object::Object::unref()].