Trait pangocairo::FontExt

pub trait FontExt {
    fn get_scaled_font(&self) -> Option<ScaledFont>;
}

Trait containing all Font methods.

Implementors

Font

Required Methods

fn get_scaled_font(&self) -> Option<ScaledFont>

Gets the cairo::ScaledFont used by self. The scaled font can be referenced and kept using cairo_scaled_font_reference.

Returns

the cairo::ScaledFont used by self, or None if self is None.