Crate pangocairo

Modules

functions

Structs

Error
A generic error capable of representing various error domains (types).

Font
Font is an interface exported by fonts for use with Cairo. The actual type of the font will depend on the particular font technology Cairo was compiled to use.

FontMap
FontMap is an interface exported by font maps for use with Cairo. The actual type of the font map will depend on the particular font technology Cairo was compiled to use.

Constants

NONE_FONT

NONE_FONT_MAP

Traits

FontExt
Trait containing all Font methods.

FontMapExt
Trait containing all FontMap methods.

FontMapExtManual