Crate pangocairo

Rust PangoCairo bindings

Rust bindings and wrappers for PangoCairo, part of gtk-rs-core.

PangoCairo 1.38 is the lowest supported version for the underlying library.

Minimum supported Rust version

Currently, the minimum supported Rust version is 1.56.0.

Documentation

Rust API - Stable

Rust API - Development

GTK Installation instructions

Using

We recommend using crates from crates.io, as demonstrated here.

If you want to track the bleeding edge, use the git dependency instead:

[dependencies]
pangocairo = { git = "https://github.com/gtk-rs/gtk-rs-core.git", package = "pangocairo" }

Avoid mixing versioned and git crates like this:

[dependencies]
pangocairo = "0.13"
pangocairo = { git = "https://github.com/gtk-rs/gtk-rs-core.git", package = "pangocairo" }

See Also

cairo

glib

pango

License

pangocairo is available under the MIT License, please refer to it.

Re-exports

pub use cairo;

pub use ffi;

pub use glib;

pub use pango;

Modules

functions

prelude

Structs

Font

Font is an interface exported by fonts for use with Cairo.

FontMap

FontMap is an interface exported by font maps for use with Cairo.

Functions

context_get_resolution

Gets the resolution for the context.

context_set_font_options

Sets the font options used when rendering text with this context.

context_set_resolution

Sets the resolution for the context.

create_context

Creates a context object set up to match the current transformation and target surface of the Cairo context.

create_layout

Creates a layout object set up to match the current transformation and target surface of the Cairo context.

error_underline_path

Add a squiggly line to the current path in the specified cairo context that approximately covers the given rectangle in the style of an underline used to indicate a spelling error.

glyph_string_path

Adds the glyphs in glyphs to the current path in the specified cairo context.

layout_line_path

Adds the text in pango::LayoutLine to the current path in the specified cairo context.

layout_path

Adds the text in a pango::Layout to the current path in the specified cairo context.

show_error_underline

Draw a squiggly line in the specified cairo context that approximately covers the given rectangle in the style of an underline used to indicate a spelling error.

show_glyph_item

Draws the glyphs in glyph_item in the specified cairo context,

show_glyph_string

Draws the glyphs in glyphs in the specified cairo context.

show_layout

Draws a pango::Layout in the specified cairo context.

show_layout_line

Draws a pango::LayoutLine in the specified cairo context.

update_context

Updates a pango::Context previously created for use with Cairo to match the current transformation and target surface of a Cairo context.

update_layout

Updates the private pango::Context of a pango::Layout created with [funccreate_layout] to match the current transformation and target surface of a Cairo context.