Module pangocairo::functions

Functions

context_get_resolution
Gets the resolution for the context. See context_set_resolution()

context_set_font_options
Sets the font options used when rendering text with this context. These options override any options that update_context() derives from the target surface.

context_set_resolution
Sets the resolution for the context. This is a scale factor between points specified in a PangoFontDescription and Cairo units. The default value is 96, meaning that a 10 point font will be 13 units high. (10 * 96. / 72. = 13.3).

create_context
Creates a context object set up to match the current transformation and target surface of the Cairo context. This context can then be used to create a layout using pango::Layout::new().

create_layout
Creates a layout object set up to match the current transformation and target surface of the Cairo context. This layout can then be used for text measurement with functions like Layout::size() or drawing with functions like show_layout(). If you change the transformation or target surface for cr, you need to call update_layout()

error_underline_path
Add a squiggly line to the current path in the specified cairo context that approximately covers the given rectangle in the style of an underline used to indicate a spelling error. (The width of the underline is rounded to an integer number of up/down segments and the resulting rectangle is centered in the original rectangle)

glyph_string_path
Adds the glyphs in glyphs to the current path in the specified cairo context. The origin of the glyphs (the left edge of the baseline) will be at the current point of the cairo context.

layout_line_path
Adds the text in pango::LayoutLine to the current path in the specified cairo context. The origin of the glyphs (the left edge of the line) will be at the current point of the cairo context.

layout_path
Adds the text in a pango::Layout to the current path in the specified cairo context. The top-left corner of the pango::Layout will be at the current point of the cairo context.

show_error_underline
Draw a squiggly line in the specified cairo context that approximately covers the given rectangle in the style of an underline used to indicate a spelling error. (The width of the underline is rounded to an integer number of up/down segments and the resulting rectangle is centered in the original rectangle)

show_glyph_item
Draws the glyphs in glyph_item in the specified cairo context, embedding the text associated with the glyphs in the output if the output format supports it (PDF for example), otherwise it acts similar to show_glyph_string().

show_glyph_string
Draws the glyphs in glyphs in the specified cairo context. The origin of the glyphs (the left edge of the baseline) will be drawn at the current point of the cairo context.

show_layout
Draws a pango::Layout in the specified cairo context. The top-left corner of the pango::Layout will be drawn at the current point of the cairo context.

show_layout_line
Draws a pango::LayoutLine in the specified cairo context. The origin of the glyphs (the left edge of the line) will be drawn at the current point of the cairo context.

update_context
Updates a pango::Context previously created for use with Cairo to match the current transformation and target surface of a Cairo context. If any layouts have been created for the context, it’s necessary to call Layout::context_changed() on those layouts.

update_layout
Updates the private pango::Context of a pango::Layout created with create_layout() to match the current transformation and target surface of a Cairo context.